1.重命名mainForm 2.配置文件ini中加入MES接口API的设置
