+++ Data Transceiver Center/packages.config
@@ -1,26 +1,18 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Csv" version="2.0.93" targetFramework="net472" />
-  <package id="FluentFTP" version="47.1.0" targetFramework="net48" requireReinstallation="true" />
-  <package id="Lextm.SharpSnmpLib" version="12.5.2" targetFramework="net472" requireReinstallation="true" />
-  <package id="Microsoft.Bcl.AsyncInterfaces" version="7.0.0" targetFramework="net472" requireReinstallation="true" />
+  <package id="Csv" version="2.0.93" targetFramework="net46" />
   <package id="Microsoft.DotNet.PlatformAbstractions" version="3.1.6" targetFramework="net48" />
-  <package id="Microsoft.Extensions.DependencyModel" version="7.0.0" targetFramework="net472" requireReinstallation="true" />
   <package id="Newtonsoft.Json" version="13.0.3" targetFramework="net48" />
   <package id="Portable.BouncyCastle" version="1.9.0" targetFramework="net48" />
-  <package id="System.Buffers" version="4.5.1" targetFramework="net472" requireReinstallation="true" />
+  <package id="System.Buffers" version="4.5.1" targetFramework="net46" />
   <package id="System.Collections.Specialized" version="4.3.0" targetFramework="net472" />
-  <package id="System.Memory" version="4.5.5" targetFramework="net472" requireReinstallation="true" />
+  <package id="System.Memory" version="4.5.5" targetFramework="net46" />
   <package id="System.Net.NameResolution" version="4.3.0" targetFramework="net472" />
   <package id="System.Net.Sockets" version="4.3.0" targetFramework="net472" />
-  <package id="System.Numerics.Vectors" version="4.5.0" targetFramework="net472" />
-  <package id="System.Resources.Extensions" version="7.0.0" targetFramework="net48" requireReinstallation="true" />
-  <package id="System.Runtime.CompilerServices.Unsafe" version="6.0.0" targetFramework="net472" requireReinstallation="true" />
+  <package id="System.Numerics.Vectors" version="4.5.0" targetFramework="net46" />
   <package id="System.Runtime.InteropServices.RuntimeInformation" version="4.3.0" targetFramework="net472" />
-  <package id="System.Text.Encodings.Web" version="7.0.0" targetFramework="net472" requireReinstallation="true" />
-  <package id="System.Text.Json" version="7.0.3" targetFramework="net48" requireReinstallation="true" />
-  <package id="System.Threading.Tasks.Extensions" version="4.5.4" targetFramework="net472" requireReinstallation="true" />
-  <package id="System.ValueTuple" version="4.5.0" targetFramework="net472" requireReinstallation="true" />
+  <package id="System.Threading.Tasks.Extensions" version="4.5.4" targetFramework="net46" />
+  <package id="System.ValueTuple" version="4.5.0" targetFramework="net46" />
   <package id="System.Xml.XmlDocument" version="4.3.0" targetFramework="net472" />
-  <package id="ZXing.Net" version="0.15.0" targetFramework="net46" />
+  <package id="ZXing.Net" version="0.16.10" targetFramework="net46" />
 </packages>
